#38864

Sequencer on CCU to support profile reconfiguration upon pointWrite.

Related work items: #35906, #38864

+++ messaging/src/main/java/a75f/io/messaging/handler/HSReconfigHandler.kt
@@ -65,7 +65,7 @@ fun reconfigureHyperstatEquips(msgObject: JsonObject, configPoint: Point) {
             true
         )
         if (configPoint.domainName == DomainName.fanOpMode) {
-            updateFanModeCache(configPoint.equipRef, pointNewValue.asInt)
+            updateFanModeCache(configPoint.equipRef, pointNewValue.asDouble.toInt())
         }
         deviceBuilder.updateDeviceAndPoints(
             config,
